# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API para gerenciar dados da NASA APOD (Astronomy Picture of the Day)",
        "title": "AstroVista API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/apods/date-range": {
            "get": {
                "description": "Retorna as imagens astronômicas do dia dentro de um intervalo de datas especificado",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "APODs"
                ],
                "summary": "Obtém APODs por intervalo de datas",
                "parameters": [
                    {
                        "type": "string",
                        "example": "\"2023-01-01\"",
                        "description": "Data de início (formato YYYY-MM-DD)",
                        "name": "start",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "example": "\"2023-01-31\"",
                        "description": "Data de fim (formato YYYY-MM-DD)",
                        "name": "end",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.ApodsDateRangeResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/apods/search": {
            "get": {
                "description": "Busca APODs com filtros, paginação e ordenação",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "APODs"
                ],
                "summary": "Pesquisa avançada de APODs",
                "parameters": [
                    {
                        "minimum": 1,
                        "type": "integer",
                        "example": 1,
                        "description": "Número da página",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "maximum": 200,
                        "minimum": 1,
                        "type": "integer",
                        "example": 20,
                        "description": "Itens por página (1-200)",
                        "name": "perPage",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "image",
                            "video",
                            "any"
                        ],
                        "type": "string",
                        "example": "image",
                        "description": "Tipo de mídia (image, video ou any)",
                        "name": "mediaType",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "example": "nebulosa",
                        "description": "Texto para busca em título e explicação",
                        "name": "search",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "example": "2023-01-01",
                        "description": "Data inicial (formato YYYY-MM-DD)",
                        "name": "startDate",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "example": "2023-01-31",
                        "description": "Data final (formato YYYY-MM-DD)",
                        "name": "endDate",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "asc",
                            "desc"
                        ],
                        "type": "string",
                        "example": "desc",
                        "description": "Ordenação (asc ou desc)",
                        "name": "sort",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.SearchResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "handlers.Apod": {
            "type": "object",
            "properties": {
                "_id": {
                    "description": "ID do MongoDB\nexample: 507f1f77bcf86cd799439011",
                    "type": "string"
                },
                "date": {
                    "description": "Data no formato string (ex: \"1995-06-16\")\nexample: 2023-01-15\nformat: date",
                    "type": "string"
                },
                "explanation": {
                    "description": "Explicação da imagem astronômica do dia\nexample: Uma bela nebulosa capturada pelo telescópio Hubble",
                    "type": "string"
                },
                "hdurl": {
                    "description": "URL da imagem em alta definição\nexample: https://apod.nasa.gov/apod/image/2301/M31_HubbleSpitzerGendler_960.jpg\nformat: uri",
                    "type": "string"
                },
                "media_type": {
                    "description": "Tipo de mídia (imagem ou vídeo)\nexample: image\nenum: image,video",
                    "type": "string"
                },
                "service_version": {
                    "description": "Versão do serviço da API\nexample: v1",
                    "type": "string"
                },
                "title": {
                    "description": "Título da imagem astronômica do dia\nexample: Galáxia de Andrômeda",
                    "type": "string"
                },
                "url": {
                    "description": "URL da imagem em resolução padrão\nexample: https://apod.nasa.gov/apod/image/2301/M31_HubbleSpitzerGendler_960.jpg\nformat: uri",
                    "type": "string"
                }
            }
        },
        "handlers.ApodsDateRangeResponse": {
            "type": "object",
            "properties": {
                "apods": {
                    "description": "Lista de APODs",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/handlers.Apod"
                    }
                },
                "count": {
                    "description": "Número total de APODs encontrados\nexample: 7",
                    "type": "integer"
                }
            }
        },
        "handlers.SearchResponse": {
            "type": "object",
            "properties": {
                "page": {
                    "description": "Número da página atual\nexample: 1",
                    "type": "integer"
                },
                "perPage": {
                    "description": "Itens por página\nexample: 20",
                    "type": "integer"
                },
                "results": {
                    "description": "Resultados da busca",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/handlers.Apod"
                    }
                },
                "totalPages": {
                    "description": "Total de páginas disponíveis\nexample: 3",
                    "type": "integer"
                },
                "totalResults": {
                    "description": "Número total de resultados encontrados\nexample: 42",
                    "type": "integer"
                }
            }
        }
    }
}